refactor: rename board domain and directory to posts
