+++ src/main/java/com/vibecoding/demo/domain/posts/controller/PostController.java
@@ -1,20 +1,20 @@
-package com.vibecoding.demo.domain.board.controller;
+package com.vibecoding.demo.domain.posts.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/board")
+@RequestMapping("/posts")
 public class PostController {
 
     @GetMapping
     public String boardList() {
-        return "board/list";
+        return "posts/list";
     }
 
     @GetMapping("/infinite")
     public String infiniteBoardList() {
-        return "board/infinite-list";
+        return "posts/infinite-list";
     }
 }
