+++ src/main/java/com/vibecoding/demo/domain/posts/dto/PostResponse.java
@@ -1,4 +1,4 @@
-package com.vibecoding.demo.domain.board.dto;
+package com.vibecoding.demo.domain.posts.dto;
 
 import java.time.LocalDateTime;
