+++ src/main/java/com/vibecoding/demo/domain/posts/dto/PostUpdateRequest.java
@@ -1,8 +1,8 @@
-package com.vibecoding.demo.domain.board.dto;
+package com.vibecoding.demo.domain.posts.dto;
 
 import jakarta.validation.constraints.NotBlank;
 
-public record PostCreateRequest(
+public record PostUpdateRequest(
         @NotBlank(message = "제목은 필수입니다.")
         String title,
         @NotBlank(message = "내용은 필수입니다.")
