+++ src/main/resources/templates/posts/infinite-list.html
@@ -29,7 +29,7 @@
     <div class="header-section">
         <h1>게시판 (무한 스크롤)</h1>
         <div class="nav-links">
-            <a href="/board">일반 페이징으로 가기</a>
+            <a href="/posts">일반 페이징으로 가기</a>
         </div>
     </div>
