+++ src/main/resources/templates/posts/list.html
