+++ src/test/java/com/vibecoding/demo/domain/posts/service/PostServiceTest.java
@@ -1,11 +1,11 @@
-package com.vibecoding.demo.domain.board.service;
+package com.vibecoding.demo.domain.posts.service;
 
-import com.vibecoding.demo.domain.board.dto.PostCreateRequest;
-import com.vibecoding.demo.domain.board.dto.PostDetailResponse;
-import com.vibecoding.demo.domain.board.dto.PostListResponse;
-import com.vibecoding.demo.domain.board.dto.PostUpdateRequest;
-import com.vibecoding.demo.domain.board.entity.Post;
-import com.vibecoding.demo.domain.board.repository.PostRepository;
+import com.vibecoding.demo.domain.posts.dto.PostCreateRequest;
+import com.vibecoding.demo.domain.posts.dto.PostDetailResponse;
+import com.vibecoding.demo.domain.posts.dto.PostListResponse;
+import com.vibecoding.demo.domain.posts.dto.PostUpdateRequest;
+import com.vibecoding.demo.domain.posts.entity.Post;
+import com.vibecoding.demo.domain.posts.repository.PostRepository;
 import com.vibecoding.demo.domain.comment.service.CommentService;
 import com.vibecoding.demo.domain.member.entity.Member;
 import com.vibecoding.demo.domain.member.entity.Role;
